+++ crates/next-core/js/src/entry/config/next.js
@@ -1,5 +1,6 @@
 import loadConfig from "next/dist/server/config";
 import { PHASE_DEVELOPMENT_SERVER } from "next/dist/shared/lib/constants";
+import loadCustomRoutes from "next/dist/lib/load-custom-routes";
 
 const loadNextConfig = async () => {
   const nextConfig = await loadConfig(PHASE_DEVELOPMENT_SERVER, process.cwd());
